Move NodeProxyWrappingValue out of public API
