+++ community/kernel/src/main/java/org/neo4j/kernel/impl/util/ValueUtils.java
@@ -34,7 +34,6 @@ import org.neo4j.graphdb.PropertyContainer;
 import org.neo4j.graphdb.Relationship;
 import org.neo4j.graphdb.spatial.Geometry;
 import org.neo4j.graphdb.spatial.Point;
-import org.neo4j.helpers.NodeProxyWrappingNodeValue;
 import org.neo4j.helpers.RelationshipProxyWrappingEdgeValue;
 import org.neo4j.helpers.collection.Iterables;
 import org.neo4j.values.AnyValue;
